+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 
 ## [Unreleased]
 
+## [0.3.0] - 2022-04-05
+
 - add engine config.
 - add date filters.
-- add and change filters on path, info, value
-- add glob support for files, pages
+- add and change filters on path, info, value.
+- add glob support for files, pages.
 
 ## [0.2.0] - 2022-04-03
 
@@ -17,6 +19,7 @@
 
 - initial release (build, server, info, list, output, md, njk, relurl).
 
-[Unreleased]: https://github.com/ankys/homura_deno/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/ankys/homura_deno/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/ankys/homura_deno/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ankys/homura_deno/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ankys/homura_deno/releases/tag/v0.1.0